+++ src/rprocessing/mode/RLangMode.java
@@ -100,7 +100,8 @@ public class RLangMode extends Mode {
    */
   @Override
   public File[] getExampleCategoryFolders() {
-    return new File[] {new File(examplesFolder, "Basics")};
+    return new File[] {new File(examplesFolder, "Basics"), new File(examplesFolder, "Libraries"),
+        new File(examplesFolder, "reference"), new File(examplesFolder, "R Packages")};
   }
 
   /**