+++ src/test/e2e/core/function/Lights1Test.java
@@ -9,7 +9,7 @@ import test.e2e.core.E2eTestBase;
 public class Lights1Test extends E2eTestBase {
 
   public Lights1Test() {
-    coreCodeTemplate = "P3D <- \"processing.opengl.PGraphics3D\"\n\nsettings <- function() {\n    size(100, 100, P3D)\n}\n\nsetup <- function() {\n    background(0)\n    noStroke()\n}\n\ndraw <- function() {\n    # Sets the default ambient and directional light\n    lights()\n    translate(20, 50, 0)\n    sphere(30)\n    translate(60, 0, 0)\n    sphere(30)\n    saveFrame(\"%s\")\n    exit()\n}\n";
+    coreCodeTemplate = "settings <- function() {\n    size(100, 100, P3D)\n}\n\nsetup <- function() {\n    background(0)\n    noStroke()\n}\n\ndraw <- function() {\n    # Sets the default ambient and directional light\n    lights()\n    translate(20, 50, 0)\n    sphere(30)\n    translate(60, 0, 0)\n    sphere(30)\n    saveFrame(\"%s\")\n    exit()\n}\n";
     referenceURI = "https://processing.org/reference/images/lights_1.png";
   }