+++ src/test/e2e/core/function/SpotLight0Test.java
@@ -9,7 +9,7 @@ import test.e2e.core.E2eTestBase;
 public class SpotLight0Test extends E2eTestBase {
 
   public SpotLight0Test() {
-    coreCodeTemplate = "P3D <- \"processing.opengl.PGraphics3D\"\nPI <- pi\n\nsettings <- function() {\n    size(100, 100, P3D)\n}\n\ndraw <- function() {\n    background(0)\n    noStroke()\n    spotLight(51, 102, 126, 80, 20, 40, -1, 0, 0, PI/2, 2)\n    translate(20, 50, 0)\n    sphere(30)\n    saveFrame(\"%s\")\n    exit()\n}\n";
+    coreCodeTemplate = "settings <- function() {\n    size(100, 100, P3D)\n}\n\ndraw <- function() {\n    background(0)\n    noStroke()\n    spotLight(51, 102, 126, 80, 20, 40, -1, 0, 0, PI/2, 2)\n    translate(20, 50, 0)\n    sphere(30)\n    saveFrame(\"%s\")\n    exit()\n}\n";
     referenceURI = "https://processing.org/reference/images/spotLight_0.png";
   }